+++ adealer/models/__init__.py
@@ -32,3 +32,4 @@ from . import vehicle_service
 from . import report_helpers
 from . import app_update
 from . import organization
+from . import dashboard